+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/databases/TestKeyValueStoreDb.java
@@ -118,6 +118,44 @@ class TestKeyValueStoreDb extends AbstractLmdbDbTest {
         });
     }
 
+
+    @Test
+    void testGetEntryCount() {
+
+        final UID uid1 = UID.of(ByteBuffer.allocateDirect(UID.UID_ARRAY_LENGTH), 0, 0, 0, 0);
+        final UID uid2 = UID.of(ByteBuffer.allocateDirect(UID.UID_ARRAY_LENGTH), 0, 0, 0, 1);
+        final UID uid3 = UID.of(ByteBuffer.allocateDirect(UID.UID_ARRAY_LENGTH), 0, 0, 0, 2);
+
+        final KeyValueStoreKey keyValueStoreKey11 = new KeyValueStoreKey(uid1, "key11");
+        final KeyValueStoreKey keyValueStoreKey12 = new KeyValueStoreKey(uid1, "key12");
+        final KeyValueStoreKey keyValueStoreKey21 = new KeyValueStoreKey(uid2, "key21");
+        final KeyValueStoreKey keyValueStoreKey22 = new KeyValueStoreKey(uid2, "key22");
+        final KeyValueStoreKey keyValueStoreKey31 = new KeyValueStoreKey(uid3, "key31");
+
+        final ValueStoreKey valueStoreKey11 = new ValueStoreKey(11, (short) 11);
+        final ValueStoreKey valueStoreKey12 = new ValueStoreKey(12, (short) 12);
+        final ValueStoreKey valueStoreKey21 = new ValueStoreKey(21, (short) 21);
+        final ValueStoreKey valueStoreKey22 = new ValueStoreKey(22, (short) 22);
+        final ValueStoreKey valueStoreKey31 = new ValueStoreKey(31, (short) 31);
+
+        lmdbEnv.doWithWriteTxn(writeTxn -> {
+            keyValueStoreDb.put(writeTxn, keyValueStoreKey11, valueStoreKey11, false);
+            keyValueStoreDb.put(writeTxn, keyValueStoreKey12, valueStoreKey12, false);
+            keyValueStoreDb.put(writeTxn, keyValueStoreKey21, valueStoreKey21, false);
+            keyValueStoreDb.put(writeTxn, keyValueStoreKey22, valueStoreKey22, false);
+            keyValueStoreDb.put(writeTxn, keyValueStoreKey31, valueStoreKey31, false);
+
+            assertThat(keyValueStoreDb.getEntryCount(writeTxn)).isEqualTo(5);
+        });
+
+        keyValueStoreDb.logRawDatabaseContents();
+
+        lmdbEnv.doWithReadTxn(readTxn -> {
+            final long entryCount = keyValueStoreDb.getEntryCount(uid2, readTxn);
+            assertThat(entryCount).isEqualTo(2);
+        });
+    }
+
     private void putEntry(final Txn<ByteBuffer> writeTxn,
                           final long uidVal,
                           final String key,